Improved handling of feeds in edit process and get_article()
This allows get_article() to include the full feed data, rather than requiring the caller to get the feed data
from a list of feed ids. Means a knock-on change in _build_feedlist to handle the two different formats the
list of selected feeds may be in (list of feed ids, or list of feed data hashes).