Fixes for some character issues.
So much hate for character encoding issues. At this point the data going into the database is guaranteed to be utf8 html with high characters replaced
with html entities to be safe. However, pulling the data out and then converting to markdown is proving hilariously horrible (I blame the
HTML::WikiConverter::Markdown module, but it could simply be Disfavour of the Perl Gods...) - unless specific entities are forcibly converted to
safe ones, the resulting string is completely broken.

For now, there's an ever-expanding table of entities that need explicit fixing by make_markdown_body in the constructor, but there has to be a better way
to do this.