Potential fix for Issue #55
Issue #55 is caused by the default maxage being applied even when an article ID is specified. This is monumentally dumb, as a valid ID should produce an article regardless of how old it is, which feed it is in, or any other filter. This shortcircuits the filtering settings, skipping them if an ID has been specified.