Fixes for autosave and login checks
Attempts to address autosave issues reported by @tobyhoward - if the compose page is left open on a laptop, and the laptop goes into hibernate mode, the normal heartbeat from autosave will not happen and the session will have expired when the laptop is woken up. Previously, session validity was only checked when creating/editing - it had assumed the autosave heartbeat would keep the session alive. This change enables session checking in autosave features as well.