Replace year lookup with configuration option
Determining the current year from the max year with users at that year
level is unreliable. This is a naive fallback on a simplistic configured
value - ideally this should be replaced by something that compares the
current date to the known years, and works out the current year from that.